// js/app/src/components/experiment/__generated__/ExperimentTokenCostsDetailsQuery.graphql.ts
export type ExperimentTokenCostsDetailsQuery$variables = {
  experimentId: string;
};
export type ExperimentTokenCostsDetailsQuery$data = {
  readonly experiment: {
    readonly __typename: "Experiment";
    readonly costDetailSummaryEntries: ReadonlyArray<{
      readonly isPrompt: boolean;
      readonly tokenType: string;
      readonly value: {
        readonly cost: number | null;
        readonly tokens: number | null;
      };
    }>;
    readonly costSummary: {
      readonly completion: {
        readonly cost: number | null;
      };
      readonly prompt: {
        readonly cost: number | null;
      };
      readonly total: {
        readonly cost: number | null;
      };
    };
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
};
export type ExperimentTokenCostsDetailsQuery = {
  response: ExperimentTokenCostsDetailsQuery$data;
  variables: ExperimentTokenCostsDetailsQuery$variables;
};